requirements.txt
Pillow==3.3.0
numpy==1.12.1

